predict_the_future

predict_the_future

predict_the_future(f, data, start_at_nth = 1, ...)

Arguments

f
a function that takes a row-wise subset of data and returns a scalar prediction

data
a data.frame used by f to make predictions. Must contain a column named t which indicates the time of each observation.

start_at_nth
start running the predictions on the nth observation. Useful when predictions on fewer than n observations are known to be unreliable.

...
additional parameters to f

Value

a vector of predictions

Examples

predict_the_future(f = function(t, x) base::mean(x),
                   data = data.frame(t = 1:6, x = 1.3^(1:6)))

#> [1] 1.300000 1.495000 1.729000 2.010775 2.351206 2.763807